Add flag for other wikis (#2)
-Added -o --other flag:
Allows the user to use a wiki's url as last argument to search
said wiki without having to modify grok.py in text editor.
When the -o flag is in use, the last argument is the new wiki's url
and is ignored in the search string's detection.

-Modified flag's detection
Changed previous flag detection using 'any' and added a for loop
to ensure the flag's argument starts with '-'. This prevents flag's
detection in search string like 'arc-en-ciel'. I've also added full
flags to detection (--verbose, --search, etc...).

-Added use cases
Added use cases to the help section to ensure the user's understanding
of how to use grok's flags and search string. Tried to explain the -o
flag the best i could since it forces the last argument to be an url.

Ps: This is my first time using git, any contructive criticism is
welcome.

Signed-off-by: Étienne Demers <edemers@dept-info.crosemont.quebec>